Deduplicate getRenewalUrlFromPurchase
